First part of refactoring Atlas in smaller components
