+++ fdm-app/app/components/custom/atlas/atlas-controls.tsx
@@ -1,0 +1,8 @@
+import { GeolocateControl, NavigationControl } from "react-map-gl"
+
+export const Controls = (
+    <div>
+        <GeolocateControl />
+        <NavigationControl />
+    </div>
+)
