+++ fdm-app/app/components/custom/atlas/atlas.tsx
@@ -1,0 +1,1 @@
+export const ZOOM_LEVEL_FIELDS = 12
