+++ fdm-app/app/routes/farm.create.$b_id_farm.atlas.tsx
@@ -136,6 +136,7 @@ export default function Index() {
                                 fieldsAvailableUrl={
                                     loaderData.fieldsAvailableUrl
                                 }
+                                fieldsSaved={null}
                             />
                         )}
                     </ClientOnly>
